Add speech disfluencies (#7)
* hipaa compliance docs (#390)

* adds hipaa compliance docs

* update hipaa doc

* makes synthesizer methods static (#394)

* Update ivr-navigation.mdx (#395)

* simpler and less risky FILLER_PHRASES

* add simple uh disfluency

* Update autogenerated API reference (#400)

* improved make_disfluency

---------

+++ vocode/streaming/synthesizer/azure_synthesizer.py
@@ -221,7 +221,7 @@ class AzureSynthesizer(BaseSynthesizer[AzureSynthesizerConfig]):
         self,
         message: str,
         ssml: str,
-        seconds: int,
+        seconds: float,
         word_boundary_event_pool: WordBoundaryEventPool,
     ) -> str:
         events = word_boundary_event_pool.get_events_sorted()
